LayoutTests/fast/css-grid-layout/grid-item-column-row-get-set-expected.txt

Test that setting and getting grid-column and grid-row works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-column and grid-row set through CSS
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('grid-column') is "10 / 15"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('grid-column-start') is "10"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('grid-column-end') is "15"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('grid-row') is "5 / 5"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('grid-row-start') is "5"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('grid-row-end') is "5"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('grid-column') is "10 / -10"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('grid-column-start') is "10"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('grid-column-end') is "-10"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('grid-row') is "-8 / 5"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('grid-row-start') is "-8"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('grid-row-end') is "5"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('grid-column') is "span 2 / 4"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('grid-column-start') is "span 2"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('grid-column-end') is "4"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('grid-row') is "3 / span 5"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('grid-row-start') is "3"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('grid-row-end') is "span 5"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('grid-column') is "span 2 / 4"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('grid-column-start') is "span 2"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('grid-column-end') is "4"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('grid-row') is "3 / span 5"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('grid-row-start') is "3"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('grid-row-end') is "span 5"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('grid-column') is "10 / 11"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('grid-column-start') is "10"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('grid-column-end') is "11"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('grid-row') is "4 / 5"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('grid-row-start') is "4"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('grid-row-end') is "5"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('grid-column') is "auto / 1"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('grid-column-end') is "1"
PASS getComputedStyle(gridItemWithSpanCustomIdent, '').getPropertyValue('grid-column') is "1 / span first"
PASS getComputedStyle(gridItemWithSpanCustomIdent, '').getPropertyValue('grid-column-start') is "1"
PASS getComputedStyle(gridItemWithSpanCustomIdent, '').getPropertyValue('grid-column-end') is "span first"
PASS getComputedStyle(gridItemWithSpanCustomIdent, '').getPropertyValue('grid-row') is "-1 / span last"
PASS getComputedStyle(gridItemWithSpanCustomIdent, '').getPropertyValue('grid-row-start') is "-1"
PASS getComputedStyle(gridItemWithSpanCustomIdent, '').getPropertyValue('grid-row-end') is "span last"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdent, '').getPropertyValue('grid-column') is "1 / span 3 first"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdent, '').getPropertyValue('grid-column-start') is "1"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdent, '').getPropertyValue('grid-column-end') is "span 3 first"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdent, '').getPropertyValue('grid-row') is "-1 / span 2 last"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdent, '').getPropertyValue('grid-row-start') is "-1"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdent, '').getPropertyValue('grid-row-end') is "span 2 last"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('grid-column') is "thirdArea / span 1"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('grid-column-start') is "thirdArea"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('grid-column-end') is "span 1"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('grid-row') is "firstArea / span 2"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('grid-row-start') is "firstArea"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('grid-row-end') is "span 2"

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-end') is 'auto'

Test getting and setting grid-column and grid-row through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "10 / 55"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "10"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "55"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "1 / 10"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "10"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 5 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "4 / span 4"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "4"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "span 4"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "-5 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "-5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "4 / -4"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "4"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "-4"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "auto / 8"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "8"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 1 / 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "3"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "5 / span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span first / 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "3"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "5 / span last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "span last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "first / last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "nav / span last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "span last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "3 first / 2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "3 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "5 nav / span 7 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "5 nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "span 7 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 3 first / -3 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span 3 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "-3 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 2 last / -1 nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span 2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "-1 nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "5 / none"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "none"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "8 / foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "8"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "nonExistent / none"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "nonExistent"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "none"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "nonExistent / foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "nonExistent"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 3 first / none"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span 3 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "none"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span last / foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 5 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span first / span last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "span last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 5 first / span 2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span 5 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 3 first / span 7 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span 3 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "span 7 last"

Test setting grid-column and grid-row back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is '18 / 19'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-end') is '19'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is '66 / 68'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-end') is '68'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-end') is 'auto'

Test getting and setting 'inherit' grid-column and grid-row through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "1 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "gridArea / 7"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "gridArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "7"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "last / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "1 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "last / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "gridArea / 7"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "gridArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "7"
PASS successfullyParsed is true

TEST COMPLETE
